Foundation: NSCoder: fix type for de/encodeInt64

+++ Foundation/Sources/NSCoder.m
@@ -169,7 +169,7 @@
 - (void) encodeObject:(id) object forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeInt:(int) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeInteger:(NSInteger) value forKey:(NSString *) key; { SUBCLASS }
-- (void) encodeInt64:(long long) value forKey:(NSString *) key; { SUBCLASS }
+- (void) encodeInt64:(int64_t) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeInt32:(int32_t) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeFloat:(float) value forKey:(NSString *) key; { SUBCLASS }
 - (void) encodeDouble:(double) value forKey:(NSString *) key; { SUBCLASS }
@@ -185,7 +185,7 @@
 - (id) decodeObjectForKey:(NSString *) key; { return SUBCLASS; }
 - (int) decodeIntForKey:(NSString *) key; { SUBCLASS; return 0; }
 - (NSInteger) decodeIntegerForKey:(NSString *) key; { SUBCLASS; return 0; }
-- (long long) decodeInt64ForKey:(NSString *) key; { SUBCLASS; return 0; }
+- (int64_t) decodeInt64ForKey:(NSString *) key; { SUBCLASS; return 0; }
 - (int32_t) decodeInt32ForKey:(NSString *) key; { SUBCLASS; return 0; }
 - (float) decodeFloatForKey:(NSString *) key; { SUBCLASS; return 0.0; }
 - (double) decodeDoubleForKey:(NSString *) key; { SUBCLASS; return 0.0; }
